+++ release/basic/basic_kbdosa/source/help/basic_kbdosa.php
@@ -1,0 +1,20 @@
+<?php 
+  $pagename = 'Osage Basic Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+<p>
+    This keyboard layout is designed for Osage. it is based on the keyboard layout from Windows 10. 
+The Gadugi font distributed with Windows 10 supports Osage. Noto Sans Osage also supports Osage.
+</p>
+
+<h2>Desktop Keyboard Layout</h2>
+<div id='osk' data-states='default shift'>
+</div>
+
+<h2>Mobile/Tablet Keyboard Layout</h2>
+
+<div id='osk-tablet' data-states='default shift'>
+</div>