+++ release/gff/gff_amharic/source/help/gff_amharic.php
@@ -5,7 +5,7 @@
   Description:      
   Create Date:      18 Sep 2009
 
-  Modified Date:    11 Apr 2021
+  Modified Date:    19 Jul 2021
   Authors:          dyacob, mcdurdin, pbaehr
   Related Files:    
   Dependencies:     
@@ -184,6 +184,9 @@ for full details on how to type all Amharic letters, numbers and punctuation.
 <div id="VersionHistory">
 <h3>Version History</h3>
 <dl>
+  <dt>Version 2.2, 19 Jul 2021</dt>
+  <dd>Extensions for Unicode 14</dd>
+  <dd>Addition of the Abyssinica SIL (Connected)</dd>
   <dt>Version 2.1, 11 Apr 2021</dt>
   <dd>CSS Colored keys.</dd>
   <dt>Version 2.0, 1 Dec 2020</dt>