+++ release/packages/gff_gurage_and_amharic/HISTORY.md
@@ -1,5 +1,14 @@
 # Amharic (አማርኛ) & Gurage (ጉራጌ) Keyboards
 
+## 0.9 (19 Jul 2021)
+* Resync with Gurage 0.9 and Amharic 2.2 packages.
+
+## 0.8.1 (22 June 2021)
+* Resync with Gurage package to fix macOS installation issue.
+
+## 0.8 (06 Jun 2021)
+* Resync with Gurage package for Unicode 14 updates.
+
 ## 0.7.4 (11 Apr 2021)
 * Resync with Amharic package to include keyboard update.